Iterate inlining and simplification until we can do no more
Inlining can open up new opportunities for simplification. The rewriter
tries to take advantage of this by doing two passes, but there can be
chains of inlinable variables which can require an arbitrary number of
passes. Instead, we can track whether or not we've inlined a variable
this pass, and continue with further passes until we haven't found such
a variable.