Add option to aggressively inline all literals and consts
This new inlining option does two things:
- Inlines any variable which looks like a literal constant
  (`hasInitDeps` is false) and which does not appear in an lvalue
  position
- Inlines any variable marked as `const` whatsoever.

This can increase program size, if a large literal is inlined over and
over. However, the idea is exactly why the Google Closure Compiler also
always inlines constants: a (the?) major use-case of this minifier is
for demoscene shaders, which are almost certainly going to be put
through compression later anyway -- so what matters is the size after
compression, not the actual output size. And reducing the redudancy that
this inlining creates can be done by the compressor in way fewer bytes
than declaring a new variable in the program text.

https://github.com/google/closure-compiler/wiki/FAQ#closure-compiler-inlined-all-my-strings-which-made-my-code-size-bigger-why-did-it-do-that

This option isn't the default; we likely want to do more testing to see
if it actually works out that way in practise (and to make sure the
output doesn't break things!) However, when I tested it on the two
largest shaders in `tests/real` (`the_real_party...` and `lunaquatic`),
it not only reduced gzip'd size but also raw output size too. So it may
just be a win all-around anyway?

- `the_real_party...`: 17708 -> 17699 (gzipped 5181 -> 5173)
- `lunaquatic`: 7356 -> 7347 (gzipped 2745 -> 2739)