Macro names: preserve the case of the original variable names (#148)
If there's a uniform "time", we used to generate the macro "VAR_TIME". Instead, we now preserve the case and generate the macro "VAR_time".
Rationale: with the change in case, we might get conflicts in the macro names (e.g. if you had both "t" and "T"). This happens for example when the names were already minified.