Split the project in two projects: a library and the binary
This lets us reuse the Shader Minifier library more easily. The first motivation is for the WebAssembly build. Other people might also want to use the library from the .NET tool.

As a side bonus, the Checker can now be used in Release mode (the --standalone flag on the executable was breaking this).