Renamer: fix renaming bug when we minify multiple shaders (#225)
The renaming of a variable inside the code didn't always match the
global name.

As a side-effect of the name renaming (introduced in the previous
commit), similar functions in two different files won't necessarily be
minified in the same way (see inout.expected). This can cause
compression problems... however, the compression test shows that it
compresses fine with Crinkler. kkrunchy may behave differently though.