Add JSON Output Format (#508)
### **Description:**

This Pull Request introduces a new output format option: JSON.

### **Motivation:**

Currently, the various output formats in `shader-minifier` are tightly
coupled with the main program. This means that if a user requires a
specific output format that isn't already supported, they need to have
that functionality accepted and merged into the core package. This
approach isn't always the most flexible or user-friendly for everyone.

By adding a JSON output option, users will have a standardized,
machine-readable format. If their desired output format is not natively
supported, they can easily parse the JSON output and generate their
preferred format themselves. This provides greater flexibility and
empowers users to adapt the minifier's output to their specific needs
without requiring direct modifications to this repository for every new
format.

### **A Note on Implementation:**

I am not deeply familiar with F# or the .NET ecosystem, so there's a
possibility that my implementation might not be idiomatic or could be
improved. I would greatly appreciate any feedback, suggestions, or
corrections regarding the code.