Support For Multi-Dimensional Arrays (#513)
Adds support for multi-dimensional arrays of the form:

```
float x[3][4][8];
```

## Approach

Takes the simple approach of replacing the optional `size` attribute of
the declaration AST with a (potentially empty) `sizes` list.

## Outstanding Issues

### Definition condensing

The `array.frag` unit test contains a case similar to the following
(simplified for example's sake)

```
float a[5];
float b[5];
```

And expects these to be condensed to a shared one line definition

```
float a[5], b[5];
```

Notice, however, that both arrays have the same size. Turns out that, if
you change the size of b to 4, for instance, the definitions still get
condensed.

```
float a[5], b[4];
```

**But this is not valid HLSL (or GLSL, I think?).**

To fix this, I made it so that array definitions of different sizes
can't be compactified. But, this breaks the `array.frag` test, since it
has a few other cases involving arrays with unspecified sizes (i.e.
`c[]`) and arrays with sizes that are evaluated as part of the
minification process (`int size = 2+2; c[size];`).

Looking for feedback here; assuming keeping array definition-sharing in
some capacity is the desired behavior.

### Non-neighboring, Mergeable Definitions

Currently only neighboring definitions are merged. I.e.

```
float a[3];
float b[2];
float c[3];
```

Produces

```
float a[3];
float b[2];
float c[3];
```

I.e. it does not merge a and c into a shared declaration. This seems to
be a function of the way the AST transform is written.

---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1210515599774891