Support for struct methods (#514)
This PR adds basic support for parsing struct methods into the AST, and
printing them back out.

## Strategy

The strategy we choose is pretty straightforward. Previously, the AST
for a `StructOrInterfaceBlock` maintained a `Decl list` of _fields_.

We define a new algebraic type:

```
StructMember =
  | MemberVariable of Decl
  | Method of FunctionType * Stmt
```

That can be either a member variable or a method. The AST for
`StructOrInterfaceBlock` changes accordingly:

```
and StructOrInterfaceBlock = {
    // ...
    members: StructMember list
}
```

## Unsupported Features

Method renaming is not supported, though should be implementable in a
way similar to field renaming (EDIT: I actually have this done in a
follow-up).

Methods with separate definitions are not supported, i.e.

```
struct MyStruct {
  int MyFunction();
};

// ... later

int MyStruct::MyFunction() {
  return 4;
}
```

though this is valid HLSL (compilable by FXC and DXC).

Nested structs are not supported, i.e.

```
struct MyStruct {
  struct MyNestedStruct {
    // ...
  };
};
```

This is kind of a non-sequitur, but it's technically valid in HLSL. To
support this I think you'd just have to add `StructOrInterfaceBlock` to
the algebraic type for `StructMember`, and parse it, but this is
orthogonal enough that we'll save that for another day.